+++ CHANGELOG.md
@@ -134,3 +134,6 @@
 * Define `kl` operator
 * Define `Empty` patter synonym
 * Define `Layered` type family
+* Define `Progress` type alias
+* Define `Continue` patter synonym
+* Define `Interrupt` patter synonym